feat(settings): split assignment notifications into its own menu
Extract the assignment due-reminder offset toggles out of the Live
Activity settings screen into a standalone AssignmentReminderSettingsView,
reached from its own row in the Notifications section (sibling above Live
Activity), matching the Android app's separate menu.

Add an isAssignmentReminderEnabled master toggle (mirrors Android's
notifyAssignments, default on) that gates the per-offset toggles. When
off, rescheduleReminders() feeds the scheduler an empty offset set, which
cancels all pending reminders via the existing preference-change pipeline.

The notification-permission prompt moves to the new screen, since the
Live Activity screen uses ActivityKit auth, not UNUserNotificationCenter.
Reuses existing localization strings; no submodule changes.
